+++ Back_End_Spring/src/main/java/pidev/afarshop/Service/Bill/IBillService.java
@@ -8,7 +8,7 @@ public interface IBillService {
     //List<Bill> retrieveAllBills();
     Bill retrieveBillById(Long billId);
     Bill addBill(Bill bill, Long orderId);
-    float calculatePaymentAmount(Long orderId, Bill bill);
+    double calculatePaymentAmount(Long orderId, Bill bill);
     List<Bill> retrieveBillByUser();
